requirements.txt
fastapi
uvicorn
python-dotenv
gigachat
yandex-ai-studio-sdk
jinja2
requests